// packages/zudoku/src/zuplo/policy-types.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PoliciesConfigFile {
  policies?: PolicyConfigurationFragment[];
  corsPolicies?: CorsPolicyConfiguration[];
}
export interface PolicyConfigurationFragment {
  name: string;
  policyType: string;
  handler: HandlerDefinition;
  options?: {
    [k: string]: unknown;
  };
}
export interface HandlerDefinition {
  module: string;
  export: string;
  options?: {
    [k: string]: unknown;
  };
}
export interface CorsPolicyConfiguration {
  name: string;
  allowCredentials?: boolean;
  maxAge?: number;
  allowedOrigins: string[] | string;
  allowedMethods?:
    | (
        | "GET"
        | "HEAD"
        | "POST"
        | "PUT"
        | "DELETE"
        | "CONNECT"
        | "OPTIONS"
        | "TRACE"
        | "PATCH"
      )[]
    | string;
  allowedHeaders?: string[] | string;
  exposeHeaders?: string[] | string;
}